sdk_dpaa: Add macless and shared mac drivers
This patch adds the code path for macless and shared mac devices.
Only the macless (virtual interface) will be available for configuration.
The shared mac code is used as a base, where several methods were
common for both macless and shared mac.
It means that a future patch will be created to remove the shared mac portion
and let only the virtual interface code.

Signed-off-by: Radu Bulie <radu-andrei.bulie@nxp.com>