net: dpaa: add support for copper backplanes using the mtip phylib wrapper
Both the fman and sdk_fman drivers have logic to remove from
phydev->supported those link modes that aren't also supported by the
Ethernet controller, so we need to extend the DTSEC_SUPPORTED (sic)
macros to also include definitions for backplane link modes and port
modes. This makes a difference to what "ethtool fm1-mac9" sees.

Also, at least on LS1046A, U-Boot's board_ft_fman_fixup_port() will
"fix up" the device tree passed to Linux, and force a fixed-link on 10G
ports (deleting the phy-handle that in our case points to the Lynx PCS,
and thus making backplane unusable), unless we specifically use
phy-connection-type = "10gbase-kr", case in which the fixup is bypassed.

So according to U-Boot's expectation, "10gbase-kr" is what we should
use, but this value isn't properly handled in the fman_memac code:

mac_probe
-> sets priv->speed based on the phy2speed[mac_dev->phy_if] lookup table
memac_initialization
-> sets mac_dev->phy_if = PHY_INTERFACE_MODE_XGMII based on priv->speed
-> memac_init
   -> init
      -> sets IF_MODE_10G based on phy_if

We need to modify the phy2speed[] lookup table to include "10gbase-kr",
because that will help priv->speed get translated to SPEED_10000 and
mac_dev->phy_if will be set to PHY_INTERFACE_MODE_XGMII. That is key in
making sure that the IF_MODE register is programmed with the IF_MODE_10G
value (0). Otherwise, the link in the PCS will come up, but traffic will
not pass through the MAC.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>